Started creating all the cards (in board class)
Creating both types of cards in the board class. Created a helper method just to stay organized and not have all the code in the constructor. The transportation cards are done 6 for each color and 8 for rainbow. The rainbow color cards hold the value "null". We need to add checks to the claim path method for rainbow cards now. Also need to discard the cards and remove them from the players hand if the path was found. I'll do this soon. And finally the paths for "2 player" games need to be removed if a double path is claimed (won't be too hard just 2 loops most likely).